# src/main/resources/api-contract/library-api-contract.yaml
openapi: 3.0.0
servers:
  - description: API Auto Mocking
    url: https://virtserver.swaggerhub.com/bharatiyas/library-api/1.0.0
info:
  description: |
    This is API specification for Library APIs.
  version: "1.0.0"
  title: Library API
  contact:
    email: sanjayfromgomi@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /users/register:
    post:
      tags:
        - users
      summary: Register a new user to the library
      operationId: registerUser
      responses:
        '201':
          description: Register a new User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLibraryUser'
        '400':
          description: Bad Request
        '409':
          description: User already exists
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryUser'
  /users/{userId}:
    get:
      tags:
        - users
      summary: Gets the details of the user by ID. This can be searched only when user logged in and can get its own details
      description: Returns the details of the user
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User found by the supplied ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryUser'
        '404':
          description: User not found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error

    put:
      tags:
        - users
      summary: Update an existing user. It can be used to update email and phone number of the user
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: UserId to update
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateLibraryUser'
        '400':
          description: Bad Request
        '404':
          description: User not found
        '406':
          description: Method not allowed
        '409':
          description: User already exists
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryUser'

    delete:
      tags:
        - users
      summary: Deletes an User
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: UserId to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: User deleted successfully
        '404':
          description: User not found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
  /users/search:
    get:
      tags:
        - users
      summary: Search users
      description: Search users based on Name
      operationId: searchUsers
      parameters:
        - name: firstName
          in: query
          description: First name of the author to be searched
          required: true
          explode: true
          schema:
            type: string
            example: "Herbet"
        - name: lastName
          in: query
          description: Last name of the author to be searched
          required: true
          explode: true
          schema:
            type: string
            example: "Shield"
      responses:
        '200':
          description: Users list matching the name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibraryUser'
        '400':
          description: Bad Request. No search criteria supplied
        '404':
          description: No Users found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
  /users/{userId}/books:
    put:
      tags:
        - users
      summary: Issue books to the user
      operationId: issueBooks
      parameters:
        - name: userId
          in: path
          description: BookId to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueBookResponse'
        '400':
          description: Bad Request
        '404':
          description: Book not found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              properties:
                authorIds:
                  type: array
                  items:
                    type: integer
                    description: Array of Book Ids
  /users/{userId}/books/{bookId}:
    delete:
      tags:
        - users
      summary: Deletes an User
      operationId: returnBooks
      parameters:
        - name: userId
          in: path
          description: UserId to delete
          required: true
          schema:
            type: integer
            format: int64
        - name: bookId
          in: path
          description: UserId to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: User deleted successfully
        '404':
          description: User not found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        bookId:
          type: integer
          example: 2001
        isbn:
          type: string
          example: "978-3-16-148410-1"
        title:
          type: string
          example: "APIs are fun"
        publisherId:
          type: integer
          example: 3476
        yearPublished:
          type: string
          example: "2015-02-22"
        edition:
          type: string
          example: "First edition"
        bookStatus:
          $ref: '#/components/schemas/BookStatus'
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
    BookStatus:
      type: object
      properties:
        bookId:
          type: integer
        state:
          enum:
            - Active
            - Inactive
        totalNumberOfCopies:
          type: integer
          example: 3
        numberOfCopiesIssued:
          type: integer
          example: 0
...
    Author:
      type: object
      properties:
        authorId:
          type: integer
          example: 8973
        firstName:
          type: string
          example: "TestauthorFn"
        lastName:
          type: string
          example: "TestauthorLn"
        dateOfBirth:
          type: string
          format: date
          example: "2000-01-20"
        gender:
          $ref: '#/components/schemas/Gender'
    Gender:
      enum:
        - Male
        - Female
        - Undisclosed
...
    CreateLibraryUser:
      type: object
      properties:
        username:
          type: string
          example: "test.username.1"
        firstName:
          type: string
          example: "TestuserFn"
        lastName:
          type: string
          example: "TestuserLn"
        dateOfBirth:
          type: string
          format: date
          example: "1996-01-20"
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
          example: "111-222-333"
        emailId:
          type: string
          example: "test.username.1@email.usr"
    UpdateLibraryUser:
      type: object
      properties:
        password:
          type: string
          example: "myPassword"
        phoneNumber:
          type: string
          example: "111-222-333"
        emailId:
          type: string
          example: "test.username.1@email.usr"
    LibraryUser:
      type: object
      properties:
        userId:
          type: integer
          example: 7839
        username:
          type: string
          example: "test.username.1"
        password:
          type: string
          example: "myPassword"
        firstName:
          type: string
          example: "TestuserFn"
        lastName:
          type: string
          example: "TestuserLn"
        dateOfBirth:
          type: string
          format: date
          example: "1996-01-20"
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
          example: "111-222-333"
        emailId:
          type: string
          example: "test.username.1@email.usr"
        issuedBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book'
    IssueBookResponse:
      type: object
      properties:
        issueBookStatuses:
          type: array
          items:
            $ref: '#/components/schemas/IssueBookStatus'
    IssueBookStatus:
      type: object
      properties:
        bookId:
          type: integer
          description: Book Id
          example: 2002
        status:
          type: string
          example: "Issued"
        remarks:
          type: string
          example: "Book Issued"